+++ custom_components/supernotify/transports/alexa_devices.py
@@ -10,6 +10,7 @@ from custom_components.supernotify import (
     OPTION_STRIP_URLS,
     OPTION_TARGET_CATEGORIES,
     OPTION_TARGET_INCLUDE_RE,
+    OPTION_UNIQUE_TARGETS,
     TRANSPORT_ALEXA,
 )
 from custom_components.supernotify.envelope import Envelope
@@ -44,6 +45,7 @@ class AlexaDevicesTransport(Transport):
             OPTION_SIMPLIFY_TEXT: True,
             OPTION_STRIP_URLS: True,
             OPTION_MESSAGE_USAGE: MessageOnlyPolicy.STANDARD,
+            OPTION_UNIQUE_TARGETS: True,
             OPTION_TARGET_CATEGORIES: [ATTR_ENTITY_ID],
             OPTION_TARGET_INCLUDE_RE: [r"notify\.[a-z0-9_]+\_(speak|announce)", r"group\.[a-z0-9_]+"],
         }