+++ tests/supernotify/test_config_yaml.py
@@ -133,7 +133,7 @@ async def test_call_action(hass: HomeAssistant) -> None:
         "supernotify", "enquire_last_notification", None, blocking=True, return_response=True
     )
     assert notification is not None
-    assert notification["_message"] == "unit test 9484"
+    assert notification["delivered_envelopes"][0]["message"] == "unit test 9484"
     assert notification["priority"] == "medium"
 
 
@@ -346,7 +346,7 @@ async def test_template_delivery(hass: HomeAssistant) -> None:
         "supernotify", "enquire_last_notification", None, blocking=True, return_response=True
     )
     assert notification is not None
-    assert notification["_message"] == "unit test 105"
+    assert notification["delivered_envelopes"][0]["message"] == "unit test 105"
     assert notification["priority"] == "high"
 
 
@@ -381,7 +381,8 @@ async def test_delivery_and_scenario(hass: HomeAssistant) -> None:
     assert call_record == {
         "domain": "media_player",
         "action": "play_media",
-        "action_data": {"media_content_type": "sound", "media_content_id": "bell_02"},
+        "action_data": {"media_content_type": "sound",
+                        "media_content_id": "bell_02"},
         "target_data": {"entity_id": "media_player.lobby"},
     }